+++ projects/forge/src/testMod/java/site/siredvin/peripheralworks/testmod/ForgePeripheralWorksTestMod.java
@@ -23,6 +23,10 @@ public final class ForgePeripheralWorksTestMod {
         Testiarium.register(PeripheralWorksGameTests.class);
         Testiarium.register(ForgeRecipeConditionsGameTests.class);
         try {
+            if (ModList.get().isLoaded("gtceu")) {
+                Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.GTCEUPeripheralGameTests"));
+                Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.GTCEUInfoGameTests"));
+            }
             if (ModList.get().isLoaded("hostilenetworks")) {
                 Testiarium.register(Class.forName("site.siredvin.peripheralworks.testmod.NeuralNetworksGameTests"));
             }